requirements.txt
scikit-learn
opencv-python
numpy
streamlit
pillow
joblib
